Refactor api.py: extract Project / Sprint / Milestone / Task routes (#370)
Continues the api.py split. Stacks on top of #369.

- Projects CRUD + hub + team/asset management (10 endpoints)
- Milestones CRUD (4 endpoints)
- Sprints CRUD + start/complete/burndown (8 endpoints)
- Tasks CRUD + claim/complete/block + stats/next (10 endpoints)
- Task comments (3 endpoints)

Total: 35 endpoints moved → routes/projects_tasks.py.

Deliberately left in api.py:
- /tasks-ui (the SPA HTML page; lives with other UI routes)
- /agents/{name}/health (sandwiched between task routes; uses
  streaming-session machinery + audit + autonomy state directly)

api.py: 7,760 → 7,235 lines (-525, -6.8%)
Cumulative since main: 11,138 → 7,235 (-3,903 lines, -35.0%)

Tests: 1542 passed, 1 skipped (federation tests need nacl, pre-existing).
